# fitsphere-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitSphere API",
    "description": "Fitness Management Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/workouts/": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get Workouts",
        "operationId": "get_workouts_workouts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkoutResponse"
                  },
                  "type": "array",
                  "title": "Response Get Workouts Workouts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Workouts"
        ],
        "summary": "Create Workout",
        "operationId": "create_workout_workouts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/workouts/search/{keyword}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Search Workouts",
        "operationId": "search_workouts_workouts_search__keyword__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Keyword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutResponse"
                  },
                  "title": "Response Search Workouts Workouts Search  Keyword  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/{workout_id}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get Workout",
        "operationId": "get_workout_workouts__workout_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workouts"
        ],
        "summary": "Update Workout",
        "operationId": "update_workout_workouts__workout_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Delete Workout",
        "operationId": "delete_workout_workouts__workout_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Workout Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sets": {
            "type": "integer",
            "title": "Sets"
          },
          "reps": {
            "type": "integer",
            "title": "Reps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sets",
          "reps"
        ],
        "title": "ExerciseResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkoutCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "workout_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Date"
          }
        },
        "type": "object",
        "required": [
          "name",
          "duration",
          "user_id"
        ],
        "title": "WorkoutCreate"
      },
      "WorkoutResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "workout_date": {
            "type": "string",
            "format": "date",
            "title": "Workout Date"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            },
            "type": "array",
            "title": "Exercises",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "duration",
          "user_id",
          "workout_date"
        ],
        "title": "WorkoutResponse"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}